Fix: Treat Unicode chars > FFFF as surrogates
Context
=======================================================================
A 'char' in .NET represents a value between 0 and 65535. This does not
conver all Unicode characters!
A unicode > 65535 has to be encoded as two chars (surrogate pair).
A surrogate is a value in the range 0xD00 - 0xDFF (reserved in unicode).

The problems
=======================================================================

1. Writer used `Encoding.UTF8.GetBytes(str)`
The encoder turns surrogate pairs into 4 byte UTF8. This is a problem
as we just went from 2 surrogate chars (str.Length = 2) to a single
UTF-8 character as far as hessian is concerned.
This fact turns string.Length useless (wrong). Therefore, we were sending
the wrong string length with the UTF8 byte stream.

2. 4 Byte UTF-8 was deserialized into a char
A char cannot hold the entire UTF8 range. It can only handle 3-byte UTF-8.
4-Byte UTF-8 would have to be turned into two surrogate chars (which
blows through buffer capacity as we now have more chars than len indicated).

The solution
=======================================================================
4 byte UTF8 is not supported in any way anymore.
We assume that surrogate pairs are used for all unicode chars
that do not fit into 3 byte UTF8.

This is the way .Net and Java handle chars internally anyway and the
way the Java lib has implemente this as well. So this implementation
is fully compatible with itself the java lib (all unicode can be sent
correctly).